Schedule page. Calendar with events.
Responsive calendar with several events scattered throughout the month of May. (Our first month of Summer Vacation)
If you stretch down your page, the calendar will adjust and be a vertical list of dates and the events and looking clean.
A note is made that a feature to click on events to learn more about them will be added.
calendar.css will be located in the CSS folder.